new file:   action.php The action.php file is used to perform the CRUD operations using PHP PDO extension and MySQL (DB class). The code is executed based on the action_type specified in the URL query string:
add – Add or insert records in the database.
edit – Update records in the database using the id given in the query string of the link.
delete – Remove data from the database using the id given in the query string of the link.
After the data manipulation, the status is stored using PHP SESSION and redirects back to the respective web page.