new file:   index.php In the index.php file, all the records are retrieved from the database and list them in tabular format with Add, Edit, and Delete links.
The Add link redirects to the add.php page to perform the Create operation.
The Edit link redirects to the edit.php page to perform the Update operation.
The Delete link redirects to the action.php file with action_type and id params. In the action.php file, the record is deleted from the database based on the row id.